Add files via upload
Not everything is implemented on both the client and server sides but lots of things have gotten started and I have tried to leave todo's and comments that explain what still need doing.

nothing tested, just code written down and theoretically working in some of them. Store on the client side is very unfinished and all around probably more end of file success/failed messages need to be sent but those I do not know as much about.