Add files via upload
Okay so the quirks of this commit are that first time around of sending RETR is that it exceptions out with a connection refused, so probably better code for waiting on the response is necessary, second time around a connection is successful but the actual reading of the file is done instantly because the readUTF() method returns an EOFException instantly because the text in the files we are trying to read from wasn't written by a writeUTF() method and thus there are some weird non printing characters that are missing from the start of each line in the files and even when i try to add them myself manually before starting the server, they still just do not take so I am not sure how to fix this part. I will be emailing the professor soon. 

Also I am sorry tensei but I still am not using tokenizer and instead am still trying to make do with resultstring.equals(response message we want) and its not going well, there was a time i was trying to use .contains(character sequence) because in m